Adds shared public-article selection for LLMS
